requirements.txt
Django>=1.9.5
django-bower
djangorestframework
mysqlclient
